Reorganize the crate
