+++ relay-profiling/src/lib.rs
@@ -112,8 +112,7 @@ mod extract_from_transaction;
 mod measurements;
 mod native_debug_image;
 mod outcomes;
-mod sample_v1;
-mod sample_v2;
+mod sample;
 mod transaction_metadata;
 mod utils;
 
@@ -130,7 +129,7 @@ struct MinimalProfile {
     event_id: ProfileId,
     platform: String,
     #[serde(default)]
-    version: sample_v1::Version,
+    version: sample::Version,
 }
 
 fn minimal_profile_from_json(
@@ -153,9 +152,9 @@ pub fn parse_metadata(payload: &[u8], project_id: ProjectId) -> Result<ProfileId
         }
     };
     match profile.version {
-        sample_v1::Version::V1 => {
+        sample::Version::V1 => {
             let d = &mut Deserializer::from_slice(payload);
-            let _: sample_v1::ProfileMetadata = match serde_path_to_error::deserialize(d) {
+            let _: sample::v1::ProfileMetadata = match serde_path_to_error::deserialize(d) {
                 Ok(profile) => profile,
                 Err(err) => {
                     relay_log::warn!(
@@ -212,8 +211,8 @@ pub fn expand_profile(payload: &[u8], event: &Event) -> Result<(ProfileId, Vec<u
     let transaction_metadata = extract_transaction_metadata(event);
     let transaction_tags = extract_transaction_tags(event);
     let processed_payload = match profile.version {
-        sample_v1::Version::V1 => {
-            sample_v1::parse_sample_profile(payload, transaction_metadata, transaction_tags)
+        sample::Version::V1 => {
+            sample::v1::parse_sample_profile(payload, transaction_metadata, transaction_tags)
         }
         _ => match profile.platform.as_str() {
             "android" => {
@@ -268,7 +267,7 @@ pub fn expand_profile_chunk(payload: &[u8]) -> Result<Vec<u8>, ProfileError> {
         }
     };
     match profile.version {
-        sample_v1::Version::V2 => sample_v2::parse(payload),
+        sample::Version::V2 => sample::v2::parse(payload),
         _ => Err(ProfileError::PlatformNotSupported),
     }
 }
@@ -282,7 +281,7 @@ mod tests {
         let data = r#"{"version":"1","platform":"cocoa","event_id":"751fff80-a266-467b-a6f5-eeeef65f4f84"}"#;
         let profile = minimal_profile_from_json(data.as_bytes());
         assert!(profile.is_ok());
-        assert_eq!(profile.unwrap().version, sample_v1::Version::V1);
+        assert_eq!(profile.unwrap().version, sample::Version::V1);
     }
 
     #[test]
@@ -290,18 +289,18 @@ mod tests {
         let data = r#"{"platform":"android","event_id":"751fff80-a266-467b-a6f5-eeeef65f4f84"}"#;
         let profile = minimal_profile_from_json(data.as_bytes());
         assert!(profile.is_ok());
-        assert_eq!(profile.unwrap().version, sample_v1::Version::Unknown);
+        assert_eq!(profile.unwrap().version, sample::Version::Unknown);
     }
 
     #[test]
     fn test_expand_profile_with_version() {
-        let payload = include_bytes!("../tests/fixtures/profiles/sample/roundtrip.json");
+        let payload = include_bytes!("../tests/fixtures/sample/v1/valid.json");
         assert!(expand_profile(payload, &Event::default()).is_ok());
     }
 
     #[test]
     fn test_expand_profile_without_version() {
-        let payload = include_bytes!("../tests/fixtures/profiles/android/roundtrip.json");
+        let payload = include_bytes!("../tests/fixtures/android/roundtrip.json");
         assert!(expand_profile(payload, &Event::default()).is_ok());
     }
 }
