+++ relay-profiling/src/sample/v1.rs
@@ -3,47 +3,17 @@ use std::ops::Range;
 
 use chrono::{DateTime, Utc};
 use itertools::Itertools;
-use relay_event_schema::protocol::{Addr, EventId, SpanId};
+use relay_event_schema::protocol::{EventId, SpanId};
 use serde::{Deserialize, Serialize};
 
 use crate::error::ProfileError;
 use crate::measurements::Measurement;
-use crate::native_debug_image::NativeDebugImage;
+use crate::sample::{DebugMeta, Frame, ThreadMetadata, Version};
 use crate::transaction_metadata::TransactionMetadata;
 use crate::utils::{deserialize_number_from_string, string_is_null_or_empty};
 use crate::MAX_PROFILE_DURATION;
 
 const MAX_PROFILE_DURATION_NS: u64 = MAX_PROFILE_DURATION.as_nanos() as u64;
-
-#[derive(Clone, Debug, Default, Serialize, Deserialize)]
-pub struct Frame {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    abs_path: Option<String>,
-    #[serde(alias = "column", skip_serializing_if = "Option::is_none")]
-    colno: Option<u32>,
-    #[serde(alias = "file", skip_serializing_if = "Option::is_none")]
-    filename: Option<String>,
-    #[serde(alias = "name", skip_serializing_if = "Option::is_none")]
-    function: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    in_app: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    instruction_addr: Option<Addr>,
-    #[serde(alias = "line", skip_serializing_if = "Option::is_none")]
-    lineno: Option<u32>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    module: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    platform: Option<String>,
-}
-
-impl Frame {
-    pub fn strip_pointer_authentication_code(&mut self, pac_code: u64) {
-        if let Some(address) = self.instruction_addr {
-            self.instruction_addr = Some(Addr(address.0 & pac_code));
-        }
-    }
-}
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 struct Sample {
@@ -56,14 +26,6 @@ struct Sample {
     // cocoa only
     #[serde(default, skip_serializing_if = "Option::is_none")]
     queue_address: Option<String>,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct ThreadMetadata {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    name: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    priority: Option<u32>,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
@@ -224,11 +186,6 @@ impl SampleProfile {
     }
 }
 
-#[derive(Default, Debug, Serialize, Deserialize, Clone)]
-pub struct DebugMeta {
-    images: Vec<NativeDebugImage>,
-}
-
 #[derive(Debug, Serialize, Deserialize, Clone)]
 struct OSMetadata {
     name: String,
@@ -256,16 +213,6 @@ struct DeviceMetadata {
     manufacturer: Option<String>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
     model: Option<String>,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone, Default, PartialEq, Eq)]
-pub enum Version {
-    #[default]
-    Unknown,
-    #[serde(rename = "1")]
-    V1,
-    #[serde(rename = "2")]
-    V2,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
@@ -416,7 +363,7 @@ mod tests {
 
     #[test]
     fn test_roundtrip() {
-        let payload = include_bytes!("../tests/fixtures/profiles/sample/roundtrip.json");
+        let payload = include_bytes!("../../tests/fixtures/sample/v1/valid.json");
         let profile = parse_profile(payload);
         assert!(profile.is_ok());
         let data = serde_json::to_vec(&profile.unwrap());
@@ -425,7 +372,7 @@ mod tests {
 
     #[test]
     fn test_expand() {
-        let payload = include_bytes!("../tests/fixtures/profiles/sample/roundtrip.json");
+        let payload = include_bytes!("../../tests/fixtures/sample/v1/valid.json");
         let profile = parse_sample_profile(payload, BTreeMap::new(), BTreeMap::new());
         assert!(profile.is_ok());
     }
@@ -921,7 +868,7 @@ mod tests {
             "some-random-transaction".to_string(),
         )]);
 
-        let payload = include_bytes!("../tests/fixtures/profiles/sample/roundtrip.json");
+        let payload = include_bytes!("../../tests/fixtures/sample/v1/valid.json");
         let profile_json = parse_sample_profile(payload, transaction_metadata, BTreeMap::new());
         assert!(profile_json.is_ok());
